MatChip code cleanup
